ML/CHATBOT/law/requirements.txt
transformers==4.36.0
sentence-transformers==2.3.1
huggingface-hub==0.20.0
langchain==0.1.0
langchain-core==0.1.16
langchain-community==0.0.13
flask==3.0.0
flask-cors==4.0.0
pypdf==3.17.0
python-dotenv==1.0.0
pinecone-client==3.0.0
groq==0.9.0
httpx==0.27.0